requirements.txt
Flask==2.3.3
requests==2.31.0
python-dotenv==1.0.0
Flask-Caching==2.1.0
Flask-CORS==4.0.0 